Allow export fonts separate from theme
This will allow the consuming app to decide how to load the theme, and how to load fonts.
