+++ Source/DesignSystem/index.ts
@@ -2,4 +2,5 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 export { componentStories, ComponentStories } from './componentStories';
-export { themeDark } from './theme';
+export { themeDark } from './theming/theme';
+
