+++ Source/DesignSystem/theming/theme.ts
@@ -6,12 +6,6 @@ import type { DataGridProComponents } from '@mui/x-data-grid-pro/themeAugmentati
 import { Components, createTheme, PaletteOptions } from '@mui/material/styles';
 import { TypographyOptions } from '@mui/material/styles/createTypography';
 
-import '@fontsource/rubik/300.css';
-import '@fontsource/rubik/400.css';
-import '@fontsource/rubik/500.css';
-import '@fontsource/rubik/700.css';
-import '@fontsource/roboto-mono';
-
 declare module '@mui/material/styles' {
     interface TypographyVariants {
         monospace: React.CSSProperties;
@@ -36,7 +30,7 @@ declare module '@mui/material/Typography' {
     }
 };
 
-const typography: TypographyOptions = {
+export const typography: TypographyOptions = {
     fontFamily: '"Rubik", "Open sans", "Arial", sans-serif',
     fontWeightMedium: 400,
     body1: {
